Move rabbit message parsers to their own module
